+++ packages/google-cloud-bigquery-analyticshub/src/v1/analytics_hub_service_client.ts
@@ -1308,7 +1308,7 @@ export class AnalyticsHubServiceClient {
    * @param {Object} request
    *   The request object that will be sent.
    * @param {google.cloud.bigquery.analyticshub.v1.DestinationDataset} request.destinationDataset
-   *   BigQuery destination dataset to create for the subscriber.
+   *   Input only. BigQuery destination dataset to create for the subscriber.
    * @param {string} request.name
    *   Required. Resource name of the listing that you want to subscribe to.
    *   e.g. `projects/myproject/locations/US/dataExchanges/123/listings/456`.
@@ -2881,7 +2881,21 @@ export class AnalyticsHubServiceClient {
    *   Required. The parent resource path of the subscription.
    *   e.g. projects/myproject/locations/US
    * @param {string} request.filter
-   *   The filter expression may be used to filter by Data Exchange or Listing.
+   *   An expression for filtering the results of the request. Eligible
+   *   fields for filtering are:
+   *
+   *    * `listing`
+   *    * `data_exchange`
+   *
+   *   Alternatively, a literal wrapped in double quotes may be provided.
+   *   This will be checked for an exact match against both fields above.
+   *
+   *   In all cases, the full Data Exchange or Listing resource name must
+   *   be provided. Some example of using filters:
+   *
+   *    * data_exchange="projects/myproject/locations/us/dataExchanges/123"
+   *    * listing="projects/123/locations/us/dataExchanges/456/listings/789"
+   *    * "projects/myproject/locations/us/dataExchanges/123"
    * @param {number} request.pageSize
    *   The maximum number of results to return in a single response page.
    * @param {string} request.pageToken
@@ -2981,7 +2995,21 @@ export class AnalyticsHubServiceClient {
    *   Required. The parent resource path of the subscription.
    *   e.g. projects/myproject/locations/US
    * @param {string} request.filter
-   *   The filter expression may be used to filter by Data Exchange or Listing.
+   *   An expression for filtering the results of the request. Eligible
+   *   fields for filtering are:
+   *
+   *    * `listing`
+   *    * `data_exchange`
+   *
+   *   Alternatively, a literal wrapped in double quotes may be provided.
+   *   This will be checked for an exact match against both fields above.
+   *
+   *   In all cases, the full Data Exchange or Listing resource name must
+   *   be provided. Some example of using filters:
+   *
+   *    * data_exchange="projects/myproject/locations/us/dataExchanges/123"
+   *    * listing="projects/123/locations/us/dataExchanges/456/listings/789"
+   *    * "projects/myproject/locations/us/dataExchanges/123"
    * @param {number} request.pageSize
    *   The maximum number of results to return in a single response page.
    * @param {string} request.pageToken
@@ -3029,7 +3057,21 @@ export class AnalyticsHubServiceClient {
    *   Required. The parent resource path of the subscription.
    *   e.g. projects/myproject/locations/US
    * @param {string} request.filter
-   *   The filter expression may be used to filter by Data Exchange or Listing.
+   *   An expression for filtering the results of the request. Eligible
+   *   fields for filtering are:
+   *
+   *    * `listing`
+   *    * `data_exchange`
+   *
+   *   Alternatively, a literal wrapped in double quotes may be provided.
+   *   This will be checked for an exact match against both fields above.
+   *
+   *   In all cases, the full Data Exchange or Listing resource name must
+   *   be provided. Some example of using filters:
+   *
+   *    * data_exchange="projects/myproject/locations/us/dataExchanges/123"
+   *    * listing="projects/123/locations/us/dataExchanges/456/listings/789"
+   *    * "projects/myproject/locations/us/dataExchanges/123"
    * @param {number} request.pageSize
    *   The maximum number of results to return in a single response page.
    * @param {string} request.pageToken